requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.2
numpy==1.26.2
tiktoken==0.5.2
google-generativeai==0.3.2
python-dotenv==1.0.0